Refactor so that tests can be added by editing test_setups.py, plus some cleanup
