+++ test_scripts/test_pypeit_test.py
@@ -25,12 +25,11 @@ To get the coverage of the tests:
 """
 
 import pytest
-
 import subprocess
 import sys
 import random
-import pypeit_test
-
+from . import test_main
+from .pypeit_tests import PypeItReduceTest
 import time
 
 
@@ -93,13 +92,13 @@ def create_dummy_files(base_path, files):
     """
     for file in files:
         full_path = base_path / file
-        full_path.parent.mkdir(parents=True)
+        full_path.parent.mkdir(parents=True, exist_ok=True)
         with open(full_path, 'w') as f:
             print("dummy content", file=f)
 
 def test_main_with_missing_file_failures(monkeypatch, tmp_path):
     """
-    Test pypeit_test.main() when there are failures due to missing files
+    Test test_main.main() when there are failures due to missing files
     """
     with monkeypatch.context() as m:
         monkeypatch.setattr(subprocess, "Popen", mock_popen)
@@ -108,16 +107,26 @@ def test_main_with_missing_file_failures(monkeypatch, tmp_path):
         # Test failure to generate pypeit file
         monkeypatch.setattr(sys, "argv", ['pypeit_test', '-o', str(tmp_path), '-t', '1',
                                           '-i', 'shane_kast_blue', '-s', '600_4310_d55', 'shane_kast_blue'])
-        assert pypeit_test.main() == 1
+        assert test_main.main() == 1
 
         # Test missing input for sensfunc
         monkeypatch.setattr(sys, "argv", ['pypeit_test', '-o', str(tmp_path), '-t', '1',
                                           '-i', 'gemini_gnirs', '-s', '32_SB_SXD', 'gemini_gnirs'])
-        assert pypeit_test.main() == 1
+        assert test_main.main() == 1
+
+        # Test not being able to find unique input for sensfunc
+        missing_files = ['gemini_gnirs/32_SB_SXD/Science/spec1d_cN20170331S0206-HIP62745_GNIRS_2017Mar31T083351.681.fits',
+                         'gemini_gnirs/32_SB_SXD/Science/spec1d_cN20170331S0206-HIP62745_GNIRS_2017Mar31T083351.682.fits']
+        create_dummy_files(tmp_path, missing_files)
+
+        monkeypatch.setattr(sys, "argv", ['pypeit_test', '-o', str(tmp_path), '-t', '1',
+                                          '-i', 'gemini_gnirs', '-s', '32_SB_SXD', 'gemini_gnirs'])
+        assert test_main.main() == 1
+
 
 def test_main_with_build_command_failure(monkeypatch, tmp_path):
     """
-    Test pypeit_test.main() when there are exceptions building commands
+    Test test_main.main() when there are exceptions building commands
     """
 
     with monkeypatch.context() as m2:
@@ -129,15 +138,15 @@ def test_main_with_build_command_failure(monkeypatch, tmp_path):
         def mock_build_command_line(self):
             raise RuntimeError("Unit testing Exception")
 
-        monkeypatch.setattr(pypeit_test.PypeItReduceTest, "build_command_line", mock_build_command_line)
+        monkeypatch.setattr(PypeItReduceTest, "build_command_line", mock_build_command_line)
         monkeypatch.setattr(sys, "argv", ['pypeit_test', '-o', str(tmp_path), '-t', '4', '-v',
                                           '-i', 'gemini_gnirs', 'develop'])
 
-        assert pypeit_test.main() == 2
+        assert test_main.main() == 2
 
 def test_main_with_test_failure(monkeypatch, tmp_path):
     """
-    Test pypeit_test.main() when a pypeit test script returns a failure.
+    Test test_main.main() when a pypeit test script returns a failure.
     """
 
     with monkeypatch.context() as m3:
@@ -150,12 +159,12 @@ def test_main_with_test_failure(monkeypatch, tmp_path):
         monkeypatch.setattr(subprocess, "run", mock_run)
 
         monkeypatch.setattr(sys, "argv", ['pypeit_test', '-o', str(tmp_path), '-t', '4', '-v', 'shane_kast_blue'])
-        assert pypeit_test.main() == 3
+        assert test_main.main() == 3
 
 
 def test_main_develop_without_failures(monkeypatch, tmp_path):
     """
-    Test pypeit_test.main() on a simulated dev suite run with no errors.
+    Test test_main.main() on a simulated dev suite run with no errors.
     """
 
     with monkeypatch.context() as m:
@@ -168,11 +177,11 @@ def test_main_develop_without_failures(monkeypatch, tmp_path):
                          'shane_kast_blue/600_4310_d55/shane_kast_blue_A/shane_kast_blue_A.pypeit',
                          'shane_kast_blue/600_4310_d55/shane_kast_blue_A/Science/spec1d_b24-Feige66_KASTb_2015May20T041246.960.fits']
         create_dummy_files(tmp_path, missing_files)
-        assert pypeit_test.main() == 0
+        assert test_main.main() == 0
 
 def test_main_debug_with_verbose_and_report(monkeypatch, tmp_path):
     """
-    Test pypeit_test.main() with the --debug option, verbose output, and an external report file
+    Test test_main.main() with the --debug option, verbose output, and an external report file
     """
     with monkeypatch.context() as m:
         monkeypatch.setattr(subprocess, "Popen", mock_popen)
@@ -182,14 +191,14 @@ def test_main_debug_with_verbose_and_report(monkeypatch, tmp_path):
         report_path = tmp_path / 'test_output.txt'
         monkeypatch.setattr(sys, "argv", ['pypeit_test', '-o', str(tmp_path), '-t', '4', '--debug',
                                           '-v', '-r', str(report_path), 'develop'])
-        assert pypeit_test.main() == 1
+        assert test_main.main() == 1
         # Verify the report exists and contains data
         stat_result = report_path.stat()
         assert stat_result.st_size > 0
 
 def test_main_with_quiet(monkeypatch, tmp_path, capsys):
     """
-    Test pypeit_test.main() with the --quiet option
+    Test test_main.main() with the --quiet option
     """
 
     with monkeypatch.context() as m:
@@ -200,7 +209,7 @@ def test_main_with_quiet(monkeypatch, tmp_path, capsys):
         # This will include a failure and skipped tests as well as passed tests
         monkeypatch.setattr(sys, "argv", ['pypeit_test', '-o', str(tmp_path), '-t', '4', '--debug',
                                           '-q', 'develop'])
-        assert pypeit_test.main() == 1
+        assert test_main.main() == 1
 
         # Verify quiet mode rally is quiet
         captured = capsys.readouterr()
@@ -214,7 +223,7 @@ def test_main_with_quiet(monkeypatch, tmp_path, capsys):
 
 def test_main_with_all(monkeypatch, tmp_path):
     """
-    Test pypeit_test.main() with the "all" argument.
+    Test test_main.main() with the "all" argument.
     """
     with monkeypatch.context() as m:
         monkeypatch.setattr(subprocess, "Popen", mock_popen)
@@ -224,12 +233,12 @@ def test_main_with_all(monkeypatch, tmp_path):
         # This will fail due to missing files
         monkeypatch.setattr(sys, "argv", ['pypeit_test', '-o', str(tmp_path), '-t', '4', 'all'])
 
-        with pytest.raises(FileNotFoundError):
-            pypeit_test.main()
+        with pytest.raises(ValueError):
+            test_main.main()
 
 def test_main_in_steps(monkeypatch, tmp_path):
     """
-    Test pypeit_test.main() prep-only, reduce, afterburn, and ql options.
+    Test test_main.main() prep-only, reduce, afterburn, and ql options.
     """
 
     with monkeypatch.context() as m:
@@ -242,16 +251,16 @@ def test_main_in_steps(monkeypatch, tmp_path):
 
         monkeypatch.setattr(sys, "argv", ['pypeit_test', '-o', str(tmp_path), '-t', '4',
                                           '-i', 'shane_kast_blue', '--prep_only', '-s', '600_4310_d55', 'develop'])
-        assert pypeit_test.main() == 0
+        assert test_main.main() == 0
 
         monkeypatch.setattr(sys, "argv", ['pypeit_test', '-o', str(tmp_path), '-t', '4',
                                           '-i', 'shane_kast_blue', '-s', '600_4310_d55', 'reduce'])
-        assert pypeit_test.main() == 0
+        assert test_main.main() == 0
 
         monkeypatch.setattr(sys, "argv", ['pypeit_test', '-o', str(tmp_path), '-t', '4',
                                           '-i', 'shane_kast_blue', '-s', '600_4310_d55', 'afterburn'])
-        assert pypeit_test.main() == 0
+        assert test_main.main() == 0
 
         monkeypatch.setattr(sys, "argv", ['pypeit_test', '-o', str(tmp_path), '-t', '4',
                                           '-i', 'shane_kast_blue', '-s', '600_4310_d55', 'ql'])
-        assert pypeit_test.main() == 0
+        assert test_main.main() == 0
